feat(cactus-web): tag component implementation
[CACTUS-339]
