+++ modules/cactus-web/src/Tag/Tag.mdx
@@ -1,0 +1,39 @@
+---
+name: Tag
+menu: Components
+---
+
+import PropsTable from 'website-src/components/PropsTable'
+import Tag from './Tag'
+
+# Tag
+
+TODO - describe component here
+
+TODO - render component here
+
+## Best practices
+
+TODO - see README for what to include here
+
+## Basic usage
+
+TODO - brief description
+
+```jsx
+import React from 'react'
+import { Tag } from '@repay/cactus-web'
+
+;<Tag />
+```
+
+```tsx
+import * as React from 'react'
+import { Tag } from '@repay/cactus-web'
+
+;<Tag />
+```
+
+## Properties
+
+<PropsTable of={Tag} />
