+++ modules/cactus-web/src/index.ts
@@ -50,3 +50,4 @@ export { ToggleField } from './ToggleField/ToggleField'
 export { Tooltip } from './Tooltip/Tooltip'
 export { default as useId } from './helpers/useId'
 export { ConfirmModal } from './ConfirmModal/ConfirmModal'
+export { Tag } from './Tag/Tag'
